Fix: Error in adding article

+++ htdocs/cashdesk/sql/llx_tmp_caisse.sql
@@ -7,7 +7,7 @@
 
 CREATE TABLE `llx_tmp_caisse` (
   `id` int(11) NOT NULL auto_increment,
-  `fk_article` tinyint(4) NOT NULL,
+  `fk_article` int(11) NOT NULL,
   `qte` int(11) NOT NULL,
   `fk_tva` int(11) NOT NULL,
   `remise_percent` int(11) NOT NULL,
